langgraph/chatbot/requirements.txt
langgraph==1.1.3
langchain-openai==1.1.12
langchain-community==0.4.1
langchain-tavily==0.2.17
python-dotenv==1.2.2
